+++ lib/Chats.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_classes_with_only_static_members
+
 import 'dart:async';
 
 import 'package:flutter/services.dart';
@@ -5,18 +7,15 @@ import 'package:flutter/services.dart';
 class Chats {
   static const MethodChannel _channel = MethodChannel('instabug_flutter');
 
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
-    return version;
-  }
+  static Future<String> get platformVersion async =>
+      (await _channel.invokeMethod<String>('getPlatformVersion'))!;
 
   @deprecated
 
   ///Use {@link BugReporting.show} instead.
   ///Manual invocation for chats view.
-  static Future<void> show() async {
-    await _channel.invokeMethod<Object>('showChats');
-  }
+  static Future<void> show() async =>
+      await _channel.invokeMethod<Object>('showChats');
 
   @deprecated