+++ lib/utils/http_client_logger.dart
@@ -4,9 +4,9 @@ import 'package:instabug_flutter/NetworkLogger.dart';
 import 'package:instabug_flutter/models/network_data.dart';
 
 class HttpClientLogger {
-  Map<int, NetworkData> requests = <int, NetworkData>{};
+  final requests = <int, NetworkData>{};
 
-  NetworkData _getRequestData(int requestHashCode) {
+  NetworkData? _getRequestData(int requestHashCode) {
     if (requests[requestHashCode] != null) {
       return requests.remove(requestHashCode);
     }
@@ -14,43 +14,40 @@ class HttpClientLogger {
   }
 
   void onRequest(HttpClientRequest request, {dynamic requestBody}) {
-    final NetworkData requestData = NetworkData();
-    requestData.startTime = DateTime.now();
-    requestData.method = request.method;
-    requestData.url = request.uri.toString();
+    final requestHeaders = <String, dynamic>{};
     request.headers.forEach((String header, dynamic value) {
-      requestData.requestHeaders[header] = value[0];
+      requestHeaders[header] = value[0];
     });
-    if (requestBody != null) {
-      requestData.requestBody = requestBody;
-    }
+    final NetworkData requestData = NetworkData(
+      startTime: DateTime.now(),
+      method: request.method,
+      url: request.uri.toString(),
+      requestHeaders: requestHeaders,
+      requestBody: requestBody ?? '',
+    );
     requests[request.hashCode] = requestData;
   }
 
   void onResponse(HttpClientResponse response, HttpClientRequest request,
       {dynamic responseBody}) {
     final DateTime endTime = DateTime.now();
-    final NetworkData networkData = _getRequestData(request.hashCode);
+    final networkData = _getRequestData(request.hashCode);
 
     if (networkData == null) {
       return;
     }
 
-    networkData.status = response.statusCode;
-    networkData.duration =
-        endTime.difference(networkData.startTime).inMilliseconds;
-    if (response.headers.contentType != null) {
-      networkData.contentType = response.headers.contentType.value;
-    }
-
-    response.headers.forEach((String header, dynamic value) {
-      networkData.responseHeaders[header] = value[0];
+    final responseHeaders = networkData.responseHeaders;
+    request.headers.forEach((String header, dynamic value) {
+      responseHeaders[header] = value[0];
     });
 
-    if (responseBody != null) {
-      networkData.responseBody = responseBody;
-    }
-
-    NetworkLogger.networkLog(networkData);
+    NetworkLogger.networkLog(networkData.copyWith(
+      status: response.statusCode,
+      duration: endTime.difference(networkData.startTime).inMilliseconds,
+      contentType: response.headers.contentType?.value,
+      responseHeaders: responseHeaders,
+      responseBody: responseBody,
+    ));
   }
 }