+++ test/instabug_flutter_test.dart
@@ -4,16 +4,16 @@ import 'dart:typed_data';
 import 'dart:ui';
 
 import 'package:flutter/cupertino.dart';
-import 'package:instabug_flutter/CrashReporting.dart';
-import 'package:instabug_flutter/Instabug.dart';
-import 'package:instabug_flutter/BugReporting.dart';
-import 'package:instabug_flutter/InstabugLog.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:instabug_flutter/Surveys.dart';
-import 'package:instabug_flutter/FeatureRequests.dart';
+import 'package:instabug_flutter/BugReporting.dart';
 import 'package:instabug_flutter/Chats.dart';
+import 'package:instabug_flutter/CrashReporting.dart';
+import 'package:instabug_flutter/FeatureRequests.dart';
+import 'package:instabug_flutter/Instabug.dart';
+import 'package:instabug_flutter/InstabugLog.dart';
 import 'package:instabug_flutter/Replies.dart';
+import 'package:instabug_flutter/Surveys.dart';
 import 'package:instabug_flutter/models/crash_data.dart';
 import 'package:instabug_flutter/models/exception_data.dart';
 import 'package:stack_trace/stack_trace.dart';
@@ -21,20 +21,20 @@ import 'package:stack_trace/stack_trace.dart';
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
   final List<MethodCall> log = <MethodCall>[];
-  final appToken = '068ba9a8c3615035e163dc5f829c73be';
+  const appToken = '068ba9a8c3615035e163dc5f829c73be';
   final List<InvocationEvent> invocationEvents = <InvocationEvent>[
     InvocationEvent.floatingButton
   ];
-  final email = "s@nta.com";
-  final name = "santa";
-  String message = "Test Message";
+  const email = 's@nta.com';
+  const name = 'santa';
+  const message = 'Test Message';
   const String userAttribute = '19';
   const Map<String, String> userAttributePair = <String, String>{
     'gender': 'female'
   };
 
   setUpAll(() async {
-    MethodChannel('instabug_flutter')
+    const MethodChannel('instabug_flutter')
         .setMockMethodCallHandler((MethodCall methodCall) async {
       log.add(methodCall);
       switch (methodCall.method) {
@@ -211,7 +211,7 @@ void main() {
   test(
       'test getTags should be called with no arguments and returns list of tags',
       () async {
-    List<String> tags = await Instabug.getTags();
+    final tags = await Instabug.getTags();
     expect(log, <Matcher>[isMethodCall('getTags', arguments: null)]);
     expect(tags, ['tag1', 'tag2']);
   });
@@ -248,7 +248,7 @@ void main() {
       'test getUserAttributeForKey should be called with a string argument and return a string',
       () async {
     const String key = 'Age';
-    final String value = await Instabug.getUserAttributeForKey(key);
+    final value = await Instabug.getUserAttributeForKey(key);
     expect(log, <Matcher>[
       isMethodCall('getUserAttributeForKey:', arguments: <dynamic>[key])
     ]);
@@ -311,7 +311,7 @@ void main() {
   });
 
   test('setPrimaryColor: Test', () async {
-    Color c = const Color.fromRGBO(255, 0, 255, 1.0);
+    const c = Color.fromRGBO(255, 0, 255, 1.0);
     final List<dynamic> args = <dynamic>[c.value];
     Instabug.setPrimaryColor(c);
     expect(log, <Matcher>[
@@ -323,7 +323,7 @@ void main() {
   });
 
   test('setUserData: Test', () async {
-    String s = "This is a String";
+    const s = 'This is a String';
     final List<dynamic> args = <dynamic>[s];
     Instabug.setUserData(s);
     expect(log, <Matcher>[
@@ -335,8 +335,8 @@ void main() {
   });
 
   test('addFileAttachmentWithURL: Test', () async {
-    String filePath = "filePath";
-    String fileName = "fileName";
+    const filePath = 'filePath';
+    const fileName = 'fileName';
     final List<dynamic> args = <dynamic>[filePath, fileName];
     Instabug.addFileAttachmentWithURL(filePath, fileName);
     expect(log, <Matcher>[
@@ -348,8 +348,8 @@ void main() {
   });
 
   test('addFileAttachmentWithData: Test', () async {
-    var bdata = new Uint8List(10);
-    String fileName = "fileName";
+    final bdata = Uint8List(10);
+    const fileName = 'fileName';
     final List<dynamic> args = <dynamic>[bdata, fileName];
     Instabug.addFileAttachmentWithData(bdata, fileName);
     expect(log, <Matcher>[
@@ -386,7 +386,8 @@ void main() {
     final List<dynamic> args = <dynamic>[screenName];
     Instabug.reportScreenChange(screenName);
     expect(log, <Matcher>[
-      isMethodCall('reportScreenChange:',
+      isMethodCall(
+        'reportScreenChange:',
         arguments: args,
       )
     ]);
@@ -396,14 +397,15 @@ void main() {
     final List<dynamic> args = <dynamic>[ReproStepsMode.enabled.toString()];
     Instabug.setReproStepsMode(ReproStepsMode.enabled);
     expect(log, <Matcher>[
-      isMethodCall('setReproStepsMode:',
+      isMethodCall(
+        'setReproStepsMode:',
         arguments: args,
       )
     ]);
   });
 
   test('setBugReportingEnabled: Test', () async {
-    bool isEnabled = false;
+    const isEnabled = false;
     final List<dynamic> args = <dynamic>[isEnabled];
     BugReporting.setEnabled(isEnabled);
     expect(log, <Matcher>[
@@ -518,7 +520,7 @@ void main() {
   });
 
   test('setSurveysEnabled: Test', () async {
-    bool isEnabled = false;
+    const isEnabled = false;
     final List<dynamic> args = <dynamic>[isEnabled];
     Surveys.setEnabled(isEnabled);
     expect(log, <Matcher>[
@@ -530,7 +532,7 @@ void main() {
   });
 
   test('setAutoShowingSurveysEnabled: Test', () async {
-    bool isEnabled = false;
+    const isEnabled = false;
     final List<dynamic> args = <dynamic>[isEnabled];
     Surveys.setAutoShowingEnabled(isEnabled);
     expect(log, <Matcher>[
@@ -562,7 +564,7 @@ void main() {
   });
 
   test('setShouldShowSurveysWelcomeScreen: Test', () async {
-    bool isEnabled = false;
+    const isEnabled = false;
     final List<dynamic> args = <dynamic>[isEnabled];
     Surveys.setShouldShowWelcomeScreen(isEnabled);
     expect(log, <Matcher>[
@@ -578,12 +580,13 @@ void main() {
     expect(log, <Matcher>[
       isMethodCall(
         'showSurveysIfAvailable',
+        arguments: null,
       )
     ]);
   });
 
   test('showSurveyWithToken Test', () async {
-    String token = "token";
+    const token = 'token';
     final List<dynamic> args = <dynamic>[token];
     Surveys.showSurvey(token);
     expect(log, <Matcher>[
@@ -595,7 +598,7 @@ void main() {
   });
 
   test('hasRespondedToSurvey Test', () async {
-    String token = "token";
+    const token = 'token';
     final List<dynamic> args = <dynamic>[token];
     Surveys.hasRespondedToSurvey(token, () => () {});
     expect(log, <Matcher>[
@@ -608,11 +611,12 @@ void main() {
 
   test('showFeatureRequests Test', () async {
     FeatureRequests.show();
-    expect(log, <Matcher>[isMethodCall('showFeatureRequests')]);
+    expect(
+        log, <Matcher>[isMethodCall('showFeatureRequests', arguments: null)]);
   });
 
   test('setEmailFieldRequiredForFeatureRequests:forAction: Test', () async {
-    bool isEmailFieldRequired = false;
+    const isEmailFieldRequired = false;
     final List<dynamic> args = <dynamic>[
       isEmailFieldRequired,
       <String>[ActionType.addCommentToFeature.toString()]
@@ -629,11 +633,11 @@ void main() {
 
   test('showChats Test', () async {
     Chats.show();
-    expect(log, <Matcher>[isMethodCall('showChats')]);
+    expect(log, <Matcher>[isMethodCall('showChats', arguments: null)]);
   });
 
   test('setChatsEnabled: Test', () async {
-    bool isEnabled = false;
+    const isEnabled = false;
     final List<dynamic> args = <dynamic>[isEnabled];
     Chats.setEnabled(isEnabled);
     expect(log, <Matcher>[
@@ -645,7 +649,7 @@ void main() {
   });
 
   test('setRepliesEnabled: Test', () async {
-    bool isEnabled = false;
+    const isEnabled = false;
     final List<dynamic> args = <dynamic>[isEnabled];
     Replies.setEnabled(isEnabled);
     expect(log, <Matcher>[
@@ -658,26 +662,29 @@ void main() {
 
   test('showReplies Test', () async {
     Replies.show();
-    expect(log, <Matcher>[isMethodCall('showReplies')]);
+    expect(log, <Matcher>[isMethodCall('showReplies', arguments: null)]);
   });
 
   test('hasChats Test', () async {
     Replies.hasChats(() => () {});
-    expect(log, <Matcher>[isMethodCall('hasChats')]);
+    expect(log, <Matcher>[isMethodCall('hasChats', arguments: null)]);
   });
 
   test('setOnNewReplyReceivedCallback Test', () async {
     Replies.setOnNewReplyReceivedCallback(() => () {});
-    expect(log, <Matcher>[isMethodCall('setOnNewReplyReceivedCallback')]);
+    expect(log, <Matcher>[
+      isMethodCall('setOnNewReplyReceivedCallback', arguments: null)
+    ]);
   });
 
   test('getUnreadRepliesCount Test', () async {
     Replies.getUnreadRepliesCount(() => () {});
-    expect(log, <Matcher>[isMethodCall('getUnreadRepliesCount')]);
+    expect(
+        log, <Matcher>[isMethodCall('getUnreadRepliesCount', arguments: null)]);
   });
 
   test('setChatNotificationEnabled: Test', () async {
-    bool isEnabled = false;
+    const isEnabled = false;
     final List<dynamic> args = <dynamic>[isEnabled];
     Replies.setInAppNotificationsEnabled(isEnabled);
     expect(log, <Matcher>[
@@ -689,7 +696,7 @@ void main() {
   });
 
   test('setCrashReportingEnabled: Test', () async {
-    bool isEnabled = false;
+    const isEnabled = false;
     final List<dynamic> args = <dynamic>[isEnabled];
     CrashReporting.setEnabled(isEnabled);
     expect(log, <Matcher>[
@@ -711,11 +718,11 @@ void main() {
       for (int i = 0; i < trace.frames.length; i++) {
         frames.add(ExceptionData(
             trace.frames[i].uri.toString(),
-            trace.frames[i].member,
-            trace.frames[i].line,
-            trace.frames[i].column == null ? 0 : trace.frames[i].column));
+            trace.frames[i].member!,
+            trace.frames[i].line!,
+            trace.frames[i].column == null ? 0 : trace.frames[i].column!));
       }
-      CrashData crashData = CrashData(
+      final crashData = CrashData(
           exception.toString(), Platform.operatingSystem.toString(), frames);
       final List<dynamic> args = <dynamic>[jsonEncode(crashData), handled];
       CrashReporting.reportHandledCrash(exception, stack);